Move dependency and result into utils
